Move env_ut_test to env_test, added description comment and renamed iterator test to tr to remove package clash
