+++ cmd/convox/env_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/convox/rack/client"
 	"github.com/convox/rack/test"
+	"github.com/stretchr/testify/assert"
 )
 
 // TestEnvGetAll ensures the environment of an app can be read.
@@ -255,4 +256,44 @@ func TestEnvApi(t *testing.T) {
 			Stdout:  "{}\n",
 		},
 	)
+}
+
+// TestParseEnvLine ensures that simple comments, empty lines, and key value
+// pairs, and invalid lines are correctly parsed by parseEnvLine
+func TestParseEnvLine(t *testing.T) {
+	tests := []struct {
+		line  string
+		valid bool
+		key   string
+		value string
+	}{
+		{"", true, "", ""},
+		{" ", true, "", ""},
+		{"	 ", true, "", ""},
+
+		{"#", true, "", ""},
+		{"# ", true, "", ""},
+		{"  #", true, "", ""},
+		{"	 #", true, "", ""},
+		{"# comment", true, "", ""},
+
+		{"An Invalid line", false, "", ""},
+
+		{"K=V", true, "K", "V"},
+		{"Key =value", true, "Key", "value"},
+		{"KEY = 123", true, "KEY", "123"},
+		{"k  =  292929", true, "k", "292929"},
+	}
+
+	for _, tr := range tests {
+		k, v, err := parseEnvLine(tr.line)
+		if tr.valid {
+			assert.NoError(t, err, "env line should be valid format")
+		} else {
+			assert.Error(t, err, "env line should be invalid format")
+		}
+
+		assert.Equal(t, tr.key, k, "for parsed env format key=value, invalid key returned")
+		assert.Equal(t, tr.value, v, "for parsed env format key=value, invalid value returned")
+	}
 }
